storybook-folder-structure: stories moved into there own folders
