+++ src/stories/Button/Button.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './button.scss';
+import './button.css';
 
 export interface ButtonProps {
   /** Is this the principal call to action on the page? */
