+++ src/stories/Header/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Button } from './Button';
-import './header.scss';
+import { Button } from '../Button/Button';
+import './header.css';
 
 type User = {
   name: string;
